mcp_servers/zap/requirements.txt
python-owasp-zap-v2.4==0.0.21
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
